+++ server/admin.py
@@ -1,16 +1,17 @@
 from database import User, Character
 
 
-async def get_users():
-    return_list = []
+def get_user_list():
+    # list of all Users with roles
+    result = {}
     for user in User.query().run():
-        print('id', user.id, user.name)
-    for user in User.query().run():
-        return_list.append({
-            'id': user.id,
-            'is_admin': user.is_admin,
+        id = user.get_id()
+        user_name = Character.get(id).name if id else None
+        result[id] = {
+            'user_id': id,
             'is_recruiter': user.is_recruiter,
-            'is_senior_recruiter': user.is_senior_recruiter,
-            'name': Character.get(user.id).name,
-        })
-    return {'info': return_list}
+            'is_snr_recruiter': user.is_senior_recruiter,
+            'is_admin': user.is_admin,
+            'name': user_name,
+        }
+    return result
