+++ server/auth.py
@@ -53,7 +53,7 @@ def admin_required(something):
         if current_user.is_admin or current_user.is_senior_recruiter:
             return something(*args, **kwargs)
         else:
-            return redirect(login_manager.login_view)
+            raise ForbiddenException("Insufficient privileges")
     return wrap
 
 # Decorator for roled access
@@ -63,7 +63,7 @@ def roles_required(something):
         if not current_user.is_applicant:
             return something(*args, **kwargs)
         else:
-            return redirect(login_manager.login_view)
+            raise ForbiddenException("Insufficient roles")
     return wrap
 
 
@@ -122,7 +122,6 @@ def api_oauth_callback():
         user = User.get(character.user_id)
         login_user(user)
         if user.is_applicant:
-            print('login', character.user_id, 'is_applicant recruiter? >', user.is_recruiter)
             return redirect(f'{react_app_url}?showing=applicant')
         return redirect(recruiter_url)
     elif login_type == 'link':
