+++ server/main.py
@@ -32,9 +32,9 @@ from recruitment import (
     get_questions, get_answers, recruiter_claim_applicant,
     recruiter_release_applicant, escalate_applicant, reject_applicant,
     edit_applicant_notes, get_applicant_notes, get_applicant_list,
-    get_user_list, get_character_search_list, submit_application
+    get_character_search_list, submit_application
 )
-from admin import get_users
+from admin import get_user_list
 import asyncio
 from auth import login_manager, login, ensure_has_access, roles_required, admin_required
 from flask_login import login_required, current_user
@@ -646,30 +646,4 @@ def api_user_answers(user_id=None):
     ensure_has_access(current_user.get_id(), user_id, self_access=True)
     return jsonify(get_answers(user_id))
 
-
-@app.route('/api/admin/users')
-@login_required
-def api_users():
-    """
-    Get information on all registered users.
-
-    Returned data is of the form {'info': [user_1, user_2, ...]}. Each user
-    dictionary has the keys `id`, `name`, `is_admin`, `is_senior_recruiter`,
-    and `is_recruiter`.
-
-    Returns:
-        response (dict)
-
-    Error codes:
-        Forbidden (403): If logged in user is not a senior recruiter or admin.
-    """
-    return jsonify(asyncio.run(get_users()))
-
-
-@app.errorhandler(500)
-def api_server_error(e):
-    # Log the error and stacktrace.
-    print('ERROR api_server_error')
-    logging.exception('An error occurred during a request.')
-    return 'An internal error occurred.', 500
 # [END app]
